back/pyproject.toml
[tool.poetry]
name = "data-collector-back"
version = "0.1.0"
description = ""
authors = ["hakuunabatata <emachadomartins1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
openai-whisper = "^20231117"
pillow = "^10.4.0"
pytesseract = "^0.3.10"
pdf2image = "^1.17.0"
flask-cors = "^4.0.1"
boto3 = "^1.34.153"
redis = "^5.0.8"
python-dotenv = "^1.0.1"
flask-ext = "^0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
